add UserTree function

+++ log RE/src/C_ACTION.java
@@ -6,7 +6,7 @@ public class C_ACTION {
 	HashMap<String, S_ACTION> s_actions_map = new HashMap<String, S_ACTION>();
 	public long startTime;
 	public long endTime;
-	public int stayTime;
+	public long stayTime;
 	HashMap<String, Integer> pre_c_actions_map = new HashMap<String, Integer>();
 	//public String last_pre_s_action = null;
 	public int accessed_account;
