+++ log RE/src/S_ACTION.java
@@ -9,7 +9,7 @@ public class S_ACTION {
 	HashMap<String, Integer> s1_paras_map = new HashMap<String, Integer>();
 	public long startTime;
 	public long endTime;
-	public int stayTime;
+	public long stayTime;
 	HashMap<String, Integer> pre_c_s_actions_map = new HashMap<String, Integer>();
 	//public String last_pre_s_action = null;
 	public int accessed_account;
